Mejora en tecnologías-lista

+++ src/app/tecnologias-lista/tecnologias-lista.component.ts
@@ -1,8 +1,11 @@
 import { Component, OnInit } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
 import { TecnologiaService } from '../services/tecnologia.service';
 
 @Component({
   selector: 'app-tecnologias-lista',
+  imports: [CommonModule, RouterModule],
   templateUrl: './tecnologias-lista.component.html',
   styleUrls: ['./tecnologias-lista.component.css']
 })
@@ -11,15 +14,22 @@ export class TecnologiasListaComponent implements OnInit {
 
   constructor(private tecnologiaService: TecnologiaService) {}
 
-  ngOnInit() {
-    this.tecnologiaService.obtenerTecnologias().subscribe(tecnologias => {
-      this.tecnologias = tecnologias;
+  ngOnInit(): void {
+    this.cargarTecnologias();
+  }
+
+  cargarTecnologias(): void {
+    this.tecnologiaService.obtenerTecnologias().subscribe((data) => {
+      this.tecnologias = data;
     });
   }
 
-  eliminarTecnologia(id: string) {
-    this.tecnologiaService.eliminarTecnologia(id).then(() => {
-      console.log('Tecnología eliminada');
-    });
+  eliminarTecnologia(id: string): void {
+    if (confirm('¿Seguro que deseas eliminar esta tecnología?')) {
+      this.tecnologiaService.eliminarTecnologia(id).then(() => {
+        alert('Tecnología eliminada correctamente.');
+        this.cargarTecnologias();
+      }).catch(error => console.error('Error eliminando tecnología:', error));
+    }
   }
 }
